fix: rename StackParamsList to StackNavigationType
