+++ App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import './gesture-handler';
 import RootProvider from './src/context';
-import AppNavigator from './src/navigation/AppNavigator';
+import AppNavigator from './src/navigation';
 
 function App(): React.JSX.Element {
   // const isDarkMode = useColorScheme() === 'dark';
