+++ src/screens/MyPageScreen.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {StackScreenProps} from '@react-navigation/stack';
 import {Text, View} from 'react-native';
-import {HomeStackParamList} from '../types/StackParamList';
+import {HomeStackParamList} from '../types/StackNavigationType';
 
 type Props = StackScreenProps<HomeStackParamList, 'MyPage'>;
